refactor(adapter-oas): split converters.ts by node family under emit/
converters.ts had grown to 769 lines, a quarter of the package in one
file. Split it into emit/converters/composition.ts (ref, allOf,
oneOf/anyOf, multi-type), scalar.ts (string, number, boolean, enum,
const, format, binary, null), and structural.ts (object, array,
tuple), plus emit/createNode.ts and emit/parseSchema.ts for the shared
context types and the ordered schemaRules dispatch table.

Part of the adapter-oas leaner-shape refactor (slice 02/07): no
behavior change, the schemaRules order and every converter body are
unchanged, only moved. Verified by the characterization snapshots.
